refactor(instance-registry): split service and http modules
